jf_vid::advz
Struct AdvzInternal

pub struct AdvzInternal<E, H, T>
where
    E: Pairing,
    T: Sync,
{ /* private fields */ }

The ADVZ VID scheme, a concrete impl for VidScheme. Consider using either Advz or AdvzGPU (enabled via gpu-vid feature).

E is any [Pairing]

H is a [digest::Digest]-compatible hash function.

T is a reference to GPU memory that’s storing SRS

Implementations

impl<E, H> AdvzInternal<E, H, ()>
where E: Pairing,

pub fn new( num_storage_nodes: u32, recovery_threshold: u32, srs: impl Borrow<<UnivariateKzgPCS<E> as PolynomialCommitmentScheme>::SRS>, ) -> VidResult<Self>

Return a new instance of Self from (mostly) implementation-independent arguments.

Implementation-independent arguments

num_storage_nodes

recovery_threshold

Implementation-specific arguments

srs

Errors

Return VidError::Argument if

num_storage_nodes < recovery_threshold

TEMPORARY recovery_threshold is not a power of two github issue

pub fn with_multiplicity( num_storage_nodes: u32, recovery_threshold: u32, max_multiplicity: u32, srs: impl Borrow<<UnivariateKzgPCS<E> as PolynomialCommitmentScheme>::SRS>, ) -> VidResult<Self>

Like Advz::new except with a max_multiplicity arg.

max_multiplicity is an implementation-specific optimization arg. Each storage node gets up to max_multiplicity evaluations per polynomial. The actual multiplicity used will be the smallest value m such that payload can fit (payload_len <= m * recovery_threshold).

Errors

In addition to Advz::new, return VidError::Argument if

TEMPORARY max_multiplicity is not a power of two github issue