jf_vid::payload_prover
Trait PayloadProver

pub trait PayloadProver<PROOF>: VidScheme
where
    PROOF: Clone + Debug + Eq + PartialEq + Serialize + DeserializeOwned,
{
    // Required methods
    fn payload_proof<B>(
        &self,
        payload: B,
        range: Range<usize>,
    ) -> VidResult<PROOF>
       where B: AsRef<[u8]>;
    fn payload_verify(
        &self,
        stmt: Statement<'_, Self>,
        proof: &PROOF,
    ) -> VidResult<Result<(), ()>>;
}

Payload proof functionality for VidScheme.

Required Methods

fn payload_proof<B>(&self, payload: B, range: Range<usize>) -> VidResult<PROOF>
where B: AsRef<[u8]>,

Compute a proof for a subslice of payload data.

Arguments

payload: a (possibly large) binary payload.

range: indicates the subslice payload[range.start..range.end] of playload for which a proof will be made.

Why not just a single &[u8] argument for payload?

You might think it’s sufficient that PayloadProver::payload_proof take only a single &[u8] argument that the user creates via payload[range.start..range.end]. However, the generated proof might depend on range.start or on payload bytes outside of range. This data would be lost if PayloadProver::payload_proof accepted only a single &[u8] argument.

fn payload_verify( &self, stmt: Statement<'_, Self>, proof: &PROOF, ) -> VidResult<Result<(), ()>>

Verify a proof made by PayloadProver::payload_proof.

Arguments

stmt: see Statement.

proof: made by a call to PayloadProver::payload_proof.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.