Trait jf_vid::precomputable::Precomputable

pub trait Precomputable: VidScheme {
    type PrecomputeData: Clone + Debug + Eq + PartialEq + Hash + Sync + Serialize + DeserializeOwned;

    // Required methods
    fn commit_only_precompute<B>(
        &self,
        payload: B,
    ) -> VidResult<(Self::Commit, Self::PrecomputeData)>
       where B: AsRef<[u8]>;
    fn disperse_precompute<B>(
        &self,
        payload: B,
        data: &Self::PrecomputeData,
    ) -> VidResult<VidDisperse<Self>>
       where B: AsRef<[u8]>;
}

Allow for precomputation of certain data for VidScheme.

Required Associated Types

type PrecomputeData: Clone + Debug + Eq + PartialEq + Hash + Sync + Serialize + DeserializeOwned

Precomputed data that can be (re-)used during disperse computation

Required Methods

fn commit_only_precompute<B>( &self, payload: B, ) -> VidResult<(Self::Commit, Self::PrecomputeData)>
where B: AsRef<[u8]>,

Similar to VidScheme::commit_only but returns additional data that can be used as input to disperse_precompute for faster dispersal.

fn disperse_precompute<B>( &self, payload: B, data: &Self::PrecomputeData, ) -> VidResult<VidDisperse<Self>>
where B: AsRef<[u8]>,

Similar to VidScheme::disperse but takes as input additional data for more efficient computation and faster disersal.

Object Safety

This trait is not object safe.