jf_vid
Trait VidScheme

pub trait VidScheme {
    type Commit: Clone + Debug + Display + DeserializeOwned + Eq + PartialEq + Hash + Serialize + Sync + for<'a> TryFrom<&'a TaggedBase64> + Into<TaggedBase64>;
    type Share: Clone + Debug + DeserializeOwned + Eq + PartialEq + Hash + Serialize + Sync;
    type Common: Clone + Debug + DeserializeOwned + Eq + PartialEq + Hash + Serialize + Sync;

    // Required methods
    fn commit_only<B>(&mut self, payload: B) -> VidResult<Self::Commit>
       where B: AsRef<[u8]>;
    fn disperse<B>(&mut self, payload: B) -> VidResult<VidDisperse<Self>>
       where B: AsRef<[u8]>;
    fn verify_share(
        &self,
        share: &Self::Share,
        common: &Self::Common,
        commit: &Self::Commit,
    ) -> VidResult<Result<(), ()>>;
    fn recover_payload(
        &self,
        shares: &[Self::Share],
        common: &Self::Common,
    ) -> VidResult<Vec<u8>>;
    fn is_consistent(
        commit: &Self::Commit,
        common: &Self::Common,
    ) -> VidResult<()>;
    fn get_payload_byte_len(common: &Self::Common) -> u32;
    fn get_num_storage_nodes(common: &Self::Common) -> u32;
    fn get_multiplicity(common: &Self::Common) -> u32;
}

VID: Verifiable Information Dispersal

Required Associated Types

type Commit: Clone + Debug + Display + DeserializeOwned + Eq + PartialEq + Hash + Serialize + Sync + for<'a> TryFrom<&'a TaggedBase64> + Into<TaggedBase64>

Payload commitment.

type Share: Clone + Debug + DeserializeOwned + Eq + PartialEq + Hash + Serialize + Sync

Share-specific data sent to a storage node.

type Common: Clone + Debug + DeserializeOwned + Eq + PartialEq + Hash + Serialize + Sync

Common data sent to all storage nodes.

Required Methods

fn commit_only<B>(&mut self, payload: B) -> VidResult<Self::Commit>
where B: AsRef<[u8]>,

Compute a payload commitment

fn disperse<B>(&mut self, payload: B) -> VidResult<VidDisperse<Self>>
where B: AsRef<[u8]>,

Compute shares to send to the storage nodes

fn verify_share( &self, share: &Self::Share, common: &Self::Common, commit: &Self::Commit, ) -> VidResult<Result<(), ()>>

Verify a share. Used by both storage node and retrieval client. Why is return type a nested Result? See https://sled.rs/errors Returns:

VidResult::Err in case of actual error

VidResult::Ok(Result::Err) if verification fails

VidResult::Ok(Result::Ok) if verification succeeds

fn recover_payload( &self, shares: &[Self::Share], common: &Self::Common, ) -> VidResult<Vec<u8>>

Recover payload from shares. Do not verify shares or check recovered payload against anything.

fn is_consistent(commit: &Self::Commit, common: &Self::Common) -> VidResult<()>

Check that a VidScheme::Common is consistent with a VidScheme::Commit.

TODO conform to nested result pattern like VidScheme::verify_share. Unfortunately, VidResult<()> is more user-friently.

fn get_payload_byte_len(common: &Self::Common) -> u32

Extract the payload byte length data from a VidScheme::Common.

fn get_num_storage_nodes(common: &Self::Common) -> u32

Extract the number of storage nodes from a VidScheme::Common.

fn get_multiplicity(common: &Self::Common) -> u32

Extract the number of poly evals per share VidScheme::Common.

Object Safety

This trait is not object safe.