Crate jf_vid

Trait and implementation for a Verifiable Information Retrieval (VID). See https://arxiv.org/abs/2111.12323 section 1.3–1.4 for intro to VID semantics.

Modules

advz

Implementation of Verifiable Information Dispersal (VID) from https://eprint.iacr.org/2021/1500.

payload_prover

Trait for additional functionality in Verifiable Information Retrieval (VID) to make and verify a proof of correctness of an arbitrary sub-slice of data from a payload.

precomputable

Trait for additional functionality in Verifiable Information Retrieval (VID) for precomputation of specific data that allows for calling methods using the data to save computation for the callee.

Structs

VidDisperse

Convenience struct to aggregate disperse data.

Enums

VidError

The error type for VidScheme methods.

Traits

VidScheme

VID: Verifiable Information Dispersal

Type Aliases

VidResult

Convenience Result wrapper for VidError.