jf_vid::advz

Module payload_prover

Implementations of PayloadProver for Advz.

Two implementations:

PROOF = SmallRangeProof: Useful for small sub-slices of payload such as an individual transaction within a block. Not snark-friendly because it requires a pairing. Consists of metadata required to verify a KZG batch proof.

PROOF = LargeRangeProof: Useful for large sub-slices of payload such as a complete namespace. Snark-friendly because it does not require a pairing. Consists of metadata required to rebuild a KZG commitment.

Structs

LargeRangeProof

A proof intended for use on large payload subslices.

SmallRangeProof

A proof intended for use on small payload subslices.