jf_vid::advz
Type Alias Advz

pub type Advz<E, H> = AdvzInternal<E, H, ()>;

Normal Advz VID that’s only using CPU

Aliased Type

struct Advz<E, H> { /* private fields */ }

Implementations

impl<E, H> Advz<E, H>
where E: Pairing,

pub fn new( num_storage_nodes: u32, recovery_threshold: u32, srs: impl Borrow<<UnivariateKzgPCS<E> as PolynomialCommitmentScheme>::SRS>, ) -> VidResult<Self>

Return a new instance of Self from (mostly) implementation-independent arguments.

Implementation-independent arguments

num_storage_nodes

recovery_threshold

Implementation-specific arguments

srs

Errors

Return VidError::Argument if

num_storage_nodes < recovery_threshold

TEMPORARY recovery_threshold is not a power of two github issue

pub fn with_multiplicity( num_storage_nodes: u32, recovery_threshold: u32, max_multiplicity: u32, srs: impl Borrow<<UnivariateKzgPCS<E> as PolynomialCommitmentScheme>::SRS>, ) -> VidResult<Self>

Like Advz::new except with a max_multiplicity arg.

max_multiplicity is an implementation-specific optimization arg. Each storage node gets up to max_multiplicity evaluations per polynomial. The actual multiplicity used will be the smallest value m such that payload can fit (payload_len <= m * recovery_threshold).

Errors

In addition to Advz::new, return VidError::Argument if

TEMPORARY max_multiplicity is not a power of two github issue